# colexecjoin: optimize the initialization of the merge joiner