Whenever the merge joiner doesn't know whether the group continues into
the next batch, it appends all tuples belonging to the group currently
being built to the "buffered group". Internally, that append requires us
that we perform a deep copy of the relevant tuples from the input batch
into a scratch space since `Enqueue` method only supports full-batch
operation.

Previously, we would always allocate the scratch batch in `Init` with
maximum possible size; however, this is too much when the input data
sets are small. This commit switches the allocation to be lazy and
utilizes our `ResetMaybeReallocate` method of dynamic batch sizing. This
shows up somewhat noticeably, for example, on
`SQL/MultinodeCockroach/InterleavedSelect` benchmark (which is using the
merge join since the interleaved reader joiner has been removed).

Addresses: #62212.

Release note: None