@yuzefovich 2021-03-24
Quick benchmarks show the following:
```
name                                                    old time/op    new time/op    delta
SQL/MultinodeCockroach/InterleavedSelect/count=10-24      1.28ms ± 1%    1.27ms ± 1%   -0.72%  (p=0.043 n=10+8)
SQL/MultinodeCockroach/InterleavedSelect/count=100-24     1.59ms ± 6%    1.51ms ± 2%   -4.89%  (p=0.000 n=10+9)
SQL/MultinodeCockroach/InterleavedSelect/count=1000-24    3.05ms ± 1%    3.05ms ± 2%     ~     (p=1.000 n=10+10)

name                                                    old alloc/op   new alloc/op   delta
SQL/MultinodeCockroach/InterleavedSelect/count=10-24       269kB ±18%     198kB ± 3%  -26.26%  (p=0.000 n=10+8)
SQL/MultinodeCockroach/InterleavedSelect/count=100-24      287kB ± 2%     233kB ± 1%  -18.74%  (p=0.000 n=9+8)
SQL/MultinodeCockroach/InterleavedSelect/count=1000-24     590kB ± 1%     548kB ± 3%   -7.15%  (p=0.000 n=10+8)

name                                                    old allocs/op  new allocs/op  delta
SQL/MultinodeCockroach/InterleavedSelect/count=10-24       1.47k ± 6%     1.43k ± 1%   -3.03%  (p=0.008 n=10+8)
SQL/MultinodeCockroach/InterleavedSelect/count=100-24      1.82k ± 2%     1.81k ± 1%     ~     (p=0.056 n=9+10)
SQL/MultinodeCockroach/InterleavedSelect/count=1000-24     3.80k ± 1%     3.81k ± 1%     ~     (p=0.283 n=10+8)
```
I think the change is well-isolated, so it could be backported. Yet it doesn't offer that much performance gains, so the backport might not be justifiable. Curious to hear thoughts.